requirements.txt
acres==0.5.0
aiofiles==24.1.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
easyocr==1.7.2
etelemetry==0.3.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httplib2==0.31.0
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
looseversion==1.3.0
lxml==6.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.3.0
prov==2.1.1
puremagic==1.30
pyclipper==1.3.0.post6
pydot==4.0.1
pygame==2.6.1
PyMuPDF==1.26.4
pyparsing==3.2.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tifffile==2025.9.9
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0+cpu
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
